refactor: extract logic from main component i
